# docs/bruno/fixtures-game-api/apispec/fixtures-game-api.openapi.yaml
...
openapi: 3.0.3
info:
  title: Fixtures Game Backend API
  version: 1.0.0
  description: API contract for the no-money football prediction game backend. Use
    `FE - ...` tags for frontend work and `Internal - ...` tags for backend jobs/admin
    maintenance.
servers:
- url: http://localhost:3001
  description: Local backend
...
paths:
  "/weekly-score/calculate-all":
    post:
      tags:
      - Internal - Scoring And Manual Maintenance
      summary: Calculate all finished scores - internal job
      operationId: post_weekly_score_calculate_all
      description: 'Internal placement: scoring cron/manual local test. Purpose: calculates
        all active groups with finished fixtures. FE should only read resulting leaderboard/history
        APIs.'
      x-fe-scope: backend-internal
      x-bruno-folder: Backend - Internal Jobs - Do Not Call From FE/Scoring And Manual
        Maintenance
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CalculateAllScoresResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/weekly-score/{friendsGroupId}/calculate":
    post:
      tags:
      - Internal - Scoring And Manual Maintenance
      summary: Calculate matchweek scores - internal job
      operationId: post_weekly_score_friendsGroupId_calculate
      description: 'Internal placement: scoring worker/manual local test after results
        are finished. Purpose: calculates one group matchweek scores. FE should only
        read resulting score/history APIs.'
      x-fe-scope: backend-internal
      x-bruno-folder: Backend - Internal Jobs - Do Not Call From FE/Scoring And Manual
        Maintenance
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CalculateScoresRequest"
            example:
              matchweek: "{{matchweek}}"
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ScoreCalculationResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/weekly-score/upsert":
    post:
      tags:
      - Internal - Scoring And Manual Maintenance
      summary: Manual weekly score upsert - internal
      operationId: post_weekly_score_upsert
      description: 'Internal placement: emergency repair/local maintenance only. Purpose:
        manually inserts or updates score rows. Do not call from FE.'
      x-fe-scope: backend-internal
      x-bruno-folder: Backend - Internal Jobs - Do Not Call From FE/Scoring And Manual
        Maintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/WeeklyScoreUpsertRequest"
            example:
              weeklyScores:
              - user_id: "{{userId}}"
                friends_group_id: "{{friendsGroupId}}"
                week_number: 1
                fixtures_predicted: 1
                points_earned: 0
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/WeeklyScoreUpsertResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/weekly-score/{friendsGroupId}/leaderboard":
    get:
      tags:
      - FE - Leaderboard
      summary: Leaderboard
      operationId: get_weekly_score_friendsGroupId_leaderboard
      description: 'FE placement: standings/leaderboard screen or trophy icon. Purpose:
        returns season ranking for a friends group. Matchweek overview already embeds
        leaderboard context for the selected week.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Leaderboard
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LeaderboardResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/profiles/id":
    get:
      tags:
      - FE - Profile Settings
      summary: My profile
      operationId: get_profiles_id
      description: 'FE placement: account/settings screen. Purpose: loads current
        user profile row.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Profile Settings
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProfileResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
    put:
      tags:
      - FE - Profile Settings
      summary: Update profile
      operationId: put_profiles_id
      description: 'FE placement: account/settings save button. Purpose: updates display
        name/avatar/color/favorite team and active group preferences.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Profile Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateProfileRequest"
            example:
              display_name: "{{displayName}}"
              favorite_team: Arsenal
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MessageResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookies set by /auth/sign-in or /auth/sign-up.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
      required:
      - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
      - message
...
    ProfilePreview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        avatar_emoji:
          type: string
          nullable: true
        color_class:
          type: string
          nullable: true
      required:
      - id
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        avatar_emoji:
          type: string
          nullable: true
        color_class:
          type: string
          nullable: true
        favorite_team:
          type: string
          nullable: true
        default_friends_group_id:
          type: string
          format: uuid
          nullable: true
        last_active_friends_group_id:
          type: string
          format: uuid
          nullable: true
        points:
          type: integer
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
    ProfileResponse:
      type: object
      properties:
        data:
          "$ref": "#/components/schemas/Profile"
      required:
      - data
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          nullable: true
        avatar_emoji:
          type: string
          nullable: true
        color_class:
          type: string
          nullable: true
        favorite_team:
          type: string
          nullable: true
        default_friends_group_id:
          type: string
          format: uuid
          nullable: true
...
    WeeklyScore:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        friends_group_id:
          type: string
          format: uuid
        week_number:
          type: integer
        fixtures_predicted:
          type: integer
        exact_score_points:
          type: integer
          description: Exact scoreline bonus points. Exact scoreline is worth 2 bonus points on top of the 1 correct-result point.
        correct_result_points:
          type: integer
          description: Correct result points. One point per fixture for correctly predicting home win, away win, or draw.
        total_goals_bonus:
          type: integer
          description: Matchweek total-goals bonus. Two points for each nearest player when comparing predicted total goals across the full matchweek against actual total goals.
        red_card_bonus:
          type: integer
          description: Red-card bonus. Five points when the selected red-card fixture has a red card.
        points_earned:
          type: integer
        group_points:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LeaderboardRow:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        friends_group_id:
          type: string
          format: uuid
        points:
          type: integer
        fixtures_predicted:
          type: integer
        exact_score_points:
          type: integer
          description: Exact scoreline bonus points. Exact scoreline is worth 2 bonus points on top of the 1 correct-result point.
        correct_result_points:
          type: integer
          description: Correct result points. One point per fixture for correctly predicting home win, away win, or draw.
        total_goals_bonus:
          type: integer
          description: Matchweek total-goals bonus. Two points for each nearest player when comparing predicted total goals across the full matchweek against actual total goals.
        red_card_bonus:
          type: integer
          description: Red-card bonus. Five points when the selected red-card fixture has a red card.
        weeks_played:
          type: integer
        rank:
          type: integer
        profile:
          "$ref": "#/components/schemas/ProfilePreview"
          nullable: true
    LeaderboardResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/LeaderboardRow"
      required:
      - data
...
    CalculateScoresRequest:
      type: object
      properties:
        matchweek:
          type: string
      required:
      - matchweek
    WeeklyScoreUpsertRequest:
      type: object
      properties:
        weeklyScores:
          type: array
          items:
            "$ref": "#/components/schemas/WeeklyScore"
      required:
      - weeklyScores
    ScoreCalculationResult:
      type: object
      properties:
        rows:
          type: array
          items:
            "$ref": "#/components/schemas/WeeklyScore"
        fixturesScored:
          type: integer
        submittedUsers:
          type: integer
        friendsGroupId:
          type: string
          format: uuid
        matchweek:
          type: string
    ScoreCalculationResponse:
      type: object
      properties:
        data:
          "$ref": "#/components/schemas/ScoreCalculationResult"
      required:
      - data
    CalculateAllScoresResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/ScoreCalculationResult"
      required:
      - data
    WeeklyScoreUpsertResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/WeeklyScore"
      required:
      - data
...